Added coins and small fixes
Resizing works great. However it completely breaks the game. The game is drawn on a window and is scaled to the display size. So the boundaries need updated when the screen is resized. I did give it a try but I am so tired I almost made it worse. 
Coins cannot be collect. They are just yellow dots that have code values for there point worth.